rip out hyperopt things from strategy, add indicator populating to hyperopt
